requirements.txt
aiogram>=3.17.0
telethon>=1.38.0
sqlalchemy[asyncio]>=2.0.0
pgvector>=0.2.0
pydantic-settings>=2.0.0
google-generativeai>=0.8.0
openai>=1.0.0
redis>=5.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
